+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
-- Fix Richeditor model update behavior
+- Fix Richeditor behaviors:
+  - model update behavior
+  - Ctrl+Enter to submit data
 
 ### Removed